+++ pyproject.toml
@@ -79,7 +79,7 @@ Source = "https://github.com/ga4gh/vrs-python"
 "Bug Tracker" = "https://github.com/ga4gh/vrs-python/issues"
 
 [project.scripts]
-vrs-annotate = "ga4gh.vrs.extras.vcf_annotation:_cli"
+vrs-annotate = "ga4gh.vrs.extras.annotator.cli:_cli"
 
 [build-system]
 requires = ["setuptools>=65.3", "setuptools_scm>=8"]
@@ -199,9 +199,6 @@ exclude = [
     "ANN001",
     "ANN201",
     "ANN202",
-]
-"src/ga4gh/vrs/extras/vcf_annotation.py" = [
-    "PTH123",  # see https://github.com/ga4gh/vrs-python/issues/482
 ]
 "src/ga4gh/vrs/extras/translator.py" = [
     "ANN001",
